feat: enhance ConnectWalletModal with WatchOnlyWallet tab and refactor ImportAddress handling
